Refactored the evaluate_forecasts part as a separate file and created a notebook for timefm execution v1
