+++ notebooks/__init__.py
@@ -1,0 +1,4 @@
+import sys
+import os
+
+sys.path.append(os.path.abspath(".."))
